col_parser/requirements.txt
python-dwca-reader
rdflib
tqdm